(maint) Break out utilsHelper to own file
